feature visits partner_show and checks 'Upload Orphan List' but presence
